requirements.txt
numpy
pandas
streamlit
tensorflow
keras
pillow